+++ CHANGELOG.md
@@ -2,8 +2,26 @@
 
 All notable changes to the "pandocplay" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+**NOTE**: This extension is still under developing. Please use it carefully.
 
-## [Unreleased]
+## [0.0.1]
 
-- Initial release
+- Initial release.
+- Following options are implemented preliminary. Use of these options will be noted in README.
+  - output.text.Add
+  - output.pane
+  - output.to
+  - input.from
+  - Pandoc.path
+  - Pandoc.args
+  - Pandoc.extension
+  - Pandoc.workdir
+
+- Followings are NOT implemented yet.
+  - input.autoFindFormat
+  - output.text.Replace
+  - output.text.addFormat
+  - output.file.enable
+  - output.file.name
+  - output.file.autoAddExtension
+  - debug.message